:truck: Restructure files to suit React, fixes #9
