--- src/app/components/bob/bob.css
@@ -1,18 +1,0 @@
-.bob {
-    text-align: center;
-}
-
-.bob-header {
-    background-color: #222;
-    height: 150px;
-    padding: 20px;
-    color: white;
-}
-
-.bob-title {
-    font-size: 1.5em;
-}
-
-.bob-intro {
-    font-size: large;
-}
