+++ src/components/Router.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
-import App from "../app";
+import App from "./App";
 
 const Router = () => (
   <BrowserRouter>
