+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import Router from './app/router/Router';
+import './css/index.css';
+import Router from './components/Router';
 import registerServiceWorker from './registerServiceWorker';
 
 ReactDOM.render(<Router />, document.getElementById('root'));
